pyproject.toml
[project]
name = "gradio-imageslider"
version = "0.1.0"
description = "ImageSlider demo using Gradio"
requires-python = ">=3.8"

dependencies = [
    "gradio>=5.27.0",
    "Pillow>=10.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."] 